requirements.txt
psycopg2==2.9.5
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
easy-thumbnails==2.8.5
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
